+++ src/main/java/umc/meme/shop/domain/artist/controller/ArtistController.java
@@ -4,7 +4,6 @@ import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import umc.meme.shop.domain.artist.dto.request.ArtistProfileDto;
-import umc.meme.shop.domain.review.dto.request.UpdateReviewDto;
 import umc.meme.shop.domain.artist.service.ArtistService;
 import umc.meme.shop.global.SuccessStatus;
 import umc.meme.shop.global.response.ApiResponse;
@@ -20,12 +19,6 @@ public class ArtistController {
     public ApiResponse updateProfile(@PathVariable Long artistId, @RequestBody ArtistProfileDto profileDto){
         artistService.updateArtistProfile(artistId, profileDto);
         return ApiResponse.SuccessResponse(SuccessStatus.ARTIST_PROFILE_UPDATE);
-    }
-
-    @Operation(summary = "리뷰 관리", description = "block 상태를 통해 리뷰 공개 유무를 결정할 수 있는 API입니다.")
-    @PatchMapping("/review/{artistId}/{portfolioId}")
-    public ApiResponse updateReview(@PathVariable Long artistId, @PathVariable Long portfolioId, @RequestBody UpdateReviewDto reviewDto){
-        return ApiResponse.SuccessResponse(SuccessStatus.REVIEW_UPDATE);
     }
 
 }
