+++ src/main/java/umc/meme/shop/domain/review/service/ReviewService.java
@@ -1,0 +1,9 @@
+package umc.meme.shop.domain.review.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class ReviewService {
+}
